Refactor: move artist components to /pages/artist. Artist header now a separate component.
