+++ src/app/App.tsx
@@ -8,8 +8,8 @@ import Content from '../components/Content'
 import Home from './pages/Home'
 import Search from './pages/Search'
 import Playlist from './pages/Playlist'
-import Album from './pages/Album'
-import Artist from './pages/Artist'
+import Album from './pages/album/Album'
+import Artist from './pages/artist/Artist'
 import Profile from './pages/Profile'
 import Account from './pages/Account'
 import Download from './pages/Download'
